+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog for LTpowerAnalyzerPython
 
+## [0.0.3] - 2025-10-18
+### Added
+- Bode GainPhase complete
+- Started working on the LNAmplifier
+- Added the Utilities folder
+
 ## [0.0.2] - 2025-10-17
 ### Added
 - Added RL2000 example